refactor: theme storage improve usability
