+++ packages/storage/lib/impl/index.ts
@@ -1,0 +1,1 @@
+export * from './exampleThemeStorage';
