+++ packages/storage/lib/index.ts
@@ -1,7 +1,2 @@
-import { createStorage } from './base';
-import { exampleThemeStorage } from './exampleThemeStorage';
-import { SessionAccessLevelEnum, StorageEnum } from './enums';
-import type { BaseStorage } from './types';
-
-export { exampleThemeStorage, createStorage, StorageEnum, SessionAccessLevelEnum };
-export type { BaseStorage };
+export type { BaseStorage } from './base/types';
+export * from './impl';
